Move init command to a separate module
