+++ edge.py
@@ -10,6 +10,8 @@ import sys
 import time
 from typing import Optional
 from serde.yaml import to_yaml, from_yaml
+
+from edge.command.init import run_init
 from edge.config import EdgeConfig, GCProjectConfig, VertexConfig, SacredConfig, StorageBucketConfig, WebAppConfig
 from edge.state import EdgeState
 from edge.sacred import setup_sacred, tear_down_sacred
@@ -384,200 +386,6 @@ def init_failed():
         "See the errors above. For technical details see error log. See README for more details.",
         style="fg:ansired"
     )
-
-
-def run_init():
-    success_title = "Initialised successfully"
-    success_message = """
-What's next? We suggest you proceed with:
-
-  Commit the new vertex:edge configuration to git:
-    git add edge.yaml && git commit -m "Initialise vertex:edge"
-
-  Configure an experiment tracker (optional):
-    ./edge.py experiment-tracker init
-
-  Configure data version control:
-    ./edge.py dvc init
-
-  Train and deploy a model (see X section of the README for more details):
-    ./edge.py vertex init
-    dvc repro ...
-    ./edge.py vertex deploy
-
-Happy herding! 🐏
-        """.strip()
-    failure_title = "Initialisation failed"
-    failure_message = "See the errors above. For technical details see error log. See README for more details."
-    with TUI(
-        "Initialising vertex:edge",
-        success_title,
-        success_message,
-        failure_title,
-        failure_message
-    ) as tui:
-        with StepTUI(message="Checking your local environment", emoji="🖥️") as step:
-            with SubStepTUI("Checking gcloud version") as sub_step:
-                gcloud_version = get_gcloud_version()
-                expected_gcloud_version_string = "2021.05.21"
-                expected_gcloud_version = Version.from_string(expected_gcloud_version_string)
-                if not gcloud_version.is_at_least(expected_gcloud_version):
-                    raise EdgeException(
-                        f"We found gcloud version {str(gcloud_version)}, "
-                        f"but we require at least {str(expected_gcloud_version)}. "
-                        "Update gcloud by running `gcloud components update`."
-                    )
-
-            with SubStepTUI("Checking kubectl version") as sub_step:
-                kubectl_version = get_kubectl_version()
-                expected_kubectl_version_string = "v1.19.0"
-                expected_kubectl_version = Version.from_string(expected_kubectl_version_string)
-                if not kubectl_version.is_at_least(expected_kubectl_version):
-                    raise EdgeException(
-                        f"We found gcloud version {str(kubectl_version)}, "
-                        f"but we require at least {str(expected_kubectl_version)}. "
-                        "Please visit https://kubernetes.io/docs/tasks/tools/ for installation instructions."
-                    )
-
-            with SubStepTUI("Checking helm version") as sub_step:
-                helm_version = get_helm_version()
-                expected_helm_version_string = "v3.5.2"
-                expected_helm_version = Version.from_string(expected_helm_version_string)
-                if not helm_version.is_at_least(expected_helm_version):
-                    raise EdgeException(
-                        f"We found gcloud version {str(helm_version)}, "
-                        f"but we require at least {str(expected_helm_version)}. "
-                        "Please visit https://helm.sh/docs/intro/install/ for installation instructions."
-                    )
-
-        with StepTUI(message="Checking your GCP environment", emoji="☁️") as step:
-            with SubStepTUI(message="️Checking if you have authenticated with gcloud") as sub_step:
-                _is_authenticated, _reason = is_authenticated()
-                if not _is_authenticated:
-                    raise EdgeException(_reason)
-
-            with SubStepTUI(message="Verifying GCloud configuration") as sub_step:
-                gcloud_account = get_gcloud_account()
-                if gcloud_account is None or gcloud_account == "":
-                    raise EdgeException(
-                        "gcloud account is unset. "
-                        "Run `gcloud auth login && gcloud auth application-default login` to authenticate "
-                        "with the correct account"
-                    )
-
-                gcloud_project = get_gcloud_project()
-                if gcloud_project is None or gcloud_project == "":
-                    raise EdgeException(
-                        "gcloud project id is unset. "
-                        "Run `gcloud config set project $PROJECT_ID` to set the correct project id"
-                    )
-
-                gcloud_region = get_gcloud_region()
-                if gcloud_region is None or gcloud_region == "":
-                    raise EdgeException(
-                        "gcloud region is unset. "
-                        "Run `gcloud config set compute/region $REGION` to set the correct region"
-                    )
-
-                sub_step.update(status=TUIStatus.NEUTRAL)
-                sub_step.set_dirty()
-
-                if not questionary.confirm(f"Is this the correct GCloud account: {gcloud_account}", qmark=qmark).ask():
-                    raise EdgeException(
-                        "Run `gcloud auth login && gcloud auth application-default login` to authenticate "
-                        "with the correct account"
-                    )
-                if not questionary.confirm(f"Is this the correct project id: {gcloud_project}", qmark=qmark).ask():
-                    raise EdgeException("Run `gcloud config set project <project_id>` to set the correct project id")
-                if not questionary.confirm(f"Is this the correct region: {gcloud_region}", qmark=qmark).ask():
-                    raise EdgeException("Run `gcloud config set compute/region <region>` to set the correct region")
-
-            with SubStepTUI(f"{gcloud_region} is available on Vertex AI") as sub_step:
-                if gcloud_region not in get_gcp_regions(gcloud_project):
-                    formatted_regions = "\n      ".join(get_gcp_regions(gcloud_project))
-                    raise EdgeException(
-                        "Vertex AI only works in certain regions. "
-                        "Please choose one of the following by running `gcloud config set compute/region <region>`:\n"
-                        f"      {formatted_regions}"
-                    )
-
-            gcloud_config = GCProjectConfig(
-                project_id=gcloud_project,
-                region=gcloud_region,
-            )
-
-            with SubStepTUI(f"Checking if project '{gcloud_project}' exists") as sub_step:
-                project_exists(gcloud_project)
-
-            with SubStepTUI(f"Checking if billing is enabled for project '{gcloud_project}'") as sub_step:
-                if not is_billing_enabled(gcloud_project):
-                    raise EdgeException(
-                        f"Billing is not enabled for project '{gcloud_project}'. "
-                        f"Please enable billing for this project following these instructions "
-                        f"https://cloud.google.com/billing/docs/how-to/modify-projectBilling is not enabled "
-                        f"for project '{gcloud_project}'."
-                    )
-
-        with StepTUI(message="Initialising Google Storage and vertex:edge state file", emoji="💾") as step:
-            with SubStepTUI("Enabling Storage API") as sub_step:
-                enable_service_api("container.googleapis.com", gcloud_project)
-
-            with SubStepTUI("Configuring Google Storage bucket", status=TUIStatus.NEUTRAL) as sub_step:
-                sub_step.set_dirty()
-                storage_bucket_name = questionary.text(
-                    "Now you need to choose a name for a storage bucket that will be used for data version control, "
-                    "model assets and keeping track of the vertex:edge state\n      "
-                    "NOTE: Storage bucket names must be unique and follow certain conventions. "
-                    "Please see the following guidelines for more information "
-                    "https://cloud.google.com/storage/docs/naming-buckets."
-                    "\n      Enter Storage bucket name to use: ",
-                    qmark=qmark
-                ).ask()
-                if storage_bucket_name is None or storage_bucket_name == "":
-                    raise EdgeException("Storage bucket name is required")
-
-            storage_config = StorageBucketConfig(
-                bucket_name=storage_bucket_name,
-                dvc_store_directory="dvcstore",
-                vertex_jobs_directory="vertex",
-            )
-            storage_state = setup_storage(gcloud_project, gcloud_region, storage_bucket_name)
-
-            _state = EdgeState(
-                storage_bucket_state=storage_state
-            )
-
-            _config = EdgeConfig(
-                google_cloud_project=gcloud_config,
-                storage_bucket=storage_config,
-            )
-
-            skip_saving_state = False
-            with SubStepTUI("Checking if vertex:edge state file exists") as sub_step:
-                if EdgeState.exists(_config):
-                    sub_step.update(
-                        "The state file already exists. "
-                        "This means that vertex:edge has already been initialised using this storage bucket.",
-                        status=TUIStatus.WARNING
-                    )
-                    sub_step.set_dirty()
-                    if not questionary.confirm(
-                        f"Do you want to delete the state and start over (this action is destructive!)",
-                        qmark=qmark,
-                        default=False,
-                    ).ask():
-                        skip_saving_state = True
-
-            if skip_saving_state:
-                with SubStepTUI("Saving state file skipped", status=TUIStatus.WARNING) as sub_step:
-                    pass
-            else:
-                with SubStepTUI("Saving state file") as sub_step:
-                    _state.save(_config)
-
-        with StepTUI(message="Saving configuration", emoji="⚙️") as step:
-            with SubStepTUI("Saving configuration to edge.yaml") as sub_step:
-                _config.save("./edge.yaml")
 
 
 def webapp_handler(_config, _args):
